+++ config/main-console.php
@@ -12,6 +12,21 @@ return [
         'shop' => [
             'class'         => 'skeeks\cms\shop\components\ShopComponent',
         ],
+
+        'i18n' => [
+            'translations'  =>
+            [
+                'skeeks/shop/app' =>
+                [
+                    'class'             => 'yii\i18n\PhpMessageSource',
+                    'basePath'          => '@skeeks/cms/shop/messages',
+                    'fileMap' => [
+                        'skeeks/shop/app' => 'app.php',
+                    ],
+                ]
+            ],
+        ],
+
     ],
 
     'modules' =>